# [get/set]RgbaUnsafe, slow before faster after